+++ packages/pg-v5/test/commands/pull.js
@@ -20,13 +20,7 @@ let push
 let pull
 let emptyResponse
 
-let opts = { encoding: 'utf8', shell: true, stdio: ['ignore', 'pipe', 'inherit'] }
-
-let stdoutHandler = function (key, func) {
-  func('result')
-}
-
-let exitHandler = function (key, func) {
+let exitHandler = (key, func) => {
   func(0)
 }
 
@@ -66,94 +60,130 @@ describe('pg', () => {
       '../lib/bastion': bastion
     })[1]
 
-    sinon.stub(Math, 'random', function () {
-      return 0
-    })
+    sinon.stub(Math, 'random', () => 0)
     emptyResponse = '00'
   })
 
   afterEach(() => {
+    tunnelStub.reset()
     Math.random.restore()
   })
 
   describe('push', () => {
-    afterEach(() => {
-      delete env.PGPORT
+    const dumpOpts = {
+      env: {
+        PGSSLMODE: 'prefer',
+        ...env
+      },
+      stdio: ['pipe', 'pipe', 2],
+      encoding: 'utf8',
+      shell: true
+    }
+    const restoreOpts = {
+      env: {
+        PGPASSWORD: 'pass',
+        ...env
+      },
+      stdio: ['pipe', 'pipe', 2],
+      encoding: 'utf8',
+      shell: true
+    }
+
+    const psql = require('../../lib/psql')
+    const childProcess = require('child_process')
+
+    let spawnStub
+
+    beforeEach(() => {
+      sinon.stub(psql, 'exec').returns(Promise.resolve(emptyResponse))
+      spawnStub = sinon.stub(childProcess, 'spawn')
     })
 
-    it('pushes out a db', sinon.test(() => {
-      let psql = require('../../lib/psql')
-      sinon.stub(psql, 'exec').returns(Promise.resolve(emptyResponse))
+    afterEach(() => {
+      psql.exec.restore()
+      spawnStub.restore()
+      delete env.PGPORT
+      delete restoreOpts.env.PGPORT
+      delete dumpOpts.env.PGPORT
+    })
 
-      let cp = sinon.mock(require('child_process'))
-      let cmd = 'env PGSSLMODE=prefer pg_dump --verbose -F c -Z 0      localdb | env PGPASSWORD="pass" pg_restore --verbose --no-acl --no-owner -U jeff -h herokai.com -p 5432 -d mydb'
-      cp.expects('spawn').withExactArgs(cmd, [], opts).once().returns(
-        {
-          stdout: {
-            on: stdoutHandler
-          },
-          on: exitHandler
-        }
-      )
+    it('pushes out a db', () => {
+      const dumpFlags = ['--verbose', '-F', 'c', '-Z', '0', 'localdb']
+      const restoreFlags = ['--verbose', '-F', 'c', '--no-acl', '--no-owner', '-U', 'jeff', '-h', 'herokai.com', '-p', '5432', '-d', 'mydb']
+
+      spawnStub.withArgs('pg_dump', dumpFlags, dumpOpts).returns({
+        stdout: {
+          pipe: () => {}
+        },
+        on: exitHandler
+      })
+      spawnStub.withArgs('pg_restore', restoreFlags, restoreOpts).returns({
+        stdin: {
+          end: () => {}
+        },
+        on: exitHandler
+      })
 
       return push.run({ args: { source: 'localdb', target: 'postgres-1' }, flags: {} })
-        .then(() => cp.verify())
+        .then(() => expect(spawnStub.callCount, 'to equal', 2))
         .then(() => expect(cli.stdout, 'to equal', 'heroku-cli: Pushing localdb ---> postgres-1\nheroku-cli: Pushing complete.\n'))
         .then(() => expect(cli.stderr, 'to equal', ''))
-        .then(() => psql.exec.restore())
-    }))
+    })
 
-    it('pushes out a db using url port', sinon.test(() => {
-      let psql = require('../../lib/psql')
-      sinon.stub(psql, 'exec').returns(Promise.resolve(emptyResponse))
-      let cp = sinon.mock(require('child_process'))
-      let cmd = 'env PGSSLMODE=prefer pg_dump --verbose -F c -Z 0   -h localhost -p 5433  localdb | env PGPASSWORD="pass" pg_restore --verbose --no-acl --no-owner -U jeff -h herokai.com -p 5432 -d mydb'
+    it('pushes out a db using url port', () => {
+      const dumpFlags = ['--verbose', '-F', 'c', '-Z', '0', '-h', 'localhost', '-p', '5433', 'localdb']
+      const restoreFlags = ['--verbose', '-F', 'c', '--no-acl', '--no-owner', '-U', 'jeff', '-h', 'herokai.com', '-p', '5432', '-d', 'mydb']
 
-      cp.expects('spawn').withExactArgs(cmd, [], opts).once().returns(
-        {
-          stdout: {
-            on: stdoutHandler
-          },
-          on: exitHandler
-        }
-      )
+      spawnStub.withArgs('pg_dump', dumpFlags, dumpOpts).returns({
+        stdout: {
+          pipe: () => {}
+        },
+        on: exitHandler
+      })
+      spawnStub.withArgs('pg_restore', restoreFlags, restoreOpts).returns({
+        stdin: {
+          end: () => {}
+        },
+        on: exitHandler
+      })
 
       return push.run({ args: { source: 'postgres://localhost:5433/localdb', target: 'postgres-1' }, flags: {} })
-        .then(() => cp.verify())
+        .then(() => expect(spawnStub.callCount, 'to equal', 2))
         .then(() => expect(cli.stdout, 'to equal', 'heroku-cli: Pushing postgres://localhost:5433/localdb ---> postgres-1\nheroku-cli: Pushing complete.\n'))
         .then(() => expect(cli.stderr, 'to equal', ''))
-        .then(() => psql.exec.restore())
-    }))
+    })
 
-    it('pushes out a db using PGPORT', sinon.test(() => {
-      env.PGPORT = 5433
+    it('pushes out a db using PGPORT', () => {
+      env.PGPORT = dumpOpts.env.PGPORT = '5433'
+      restoreOpts.env.PGPORT = '5433'
 
-      let psql = require('../../lib/psql')
-      sinon.stub(psql, 'exec').returns(Promise.resolve(emptyResponse))
-      let cp = sinon.mock(require('child_process'))
+      const dumpFlags = ['--verbose', '-F', 'c', '-Z', '0', '-p', '5433', 'localdb']
+      const restoreFlags = ['--verbose', '-F', 'c', '--no-acl', '--no-owner', '-U', 'jeff', '-h', 'herokai.com', '-p', '5432', '-d', 'mydb']
 
-      let cmd = 'env PGSSLMODE=prefer pg_dump --verbose -F c -Z 0    -p 5433  localdb | env PGPASSWORD="pass" pg_restore --verbose --no-acl --no-owner -U jeff -h herokai.com -p 5432 -d mydb'
-      cp.expects('spawn').withExactArgs(cmd, [], opts).once().returns(
-        {
-          stdout: {
-            on: stdoutHandler
-          },
-          on: exitHandler
-        }
-      )
+      spawnStub.withArgs('pg_dump', dumpFlags, dumpOpts).returns({
+        stdout: {
+          pipe: () => {}
+        },
+        on: exitHandler
+      })
+      spawnStub.withArgs('pg_restore', restoreFlags, restoreOpts).returns({
+        stdin: {
+          end: () => {}
+        },
+        on: exitHandler
+      })
 
       return push.run({ args: { source: 'localdb', target: 'postgres-1' }, flags: {} })
-        .then(() => cp.verify())
+        .then(() => expect(spawnStub.callCount, 'to equal', 2))
         .then(() => expect(cli.stdout, 'to equal', 'heroku-cli: Pushing localdb ---> postgres-1\nheroku-cli: Pushing complete.\n'))
         .then(() => expect(cli.stderr, 'to equal', ''))
-        .then(() => psql.exec.restore())
-    }))
+    })
 
-    it('opens an SSH tunnel and runs pg_dump for bastion databases', sinon.test(() => {
+    it('opens an SSH tunnel and runs pg_dump for bastion databases', () => {
       db.bastionHost = 'bastion-host'
       db.bastionKey = 'super-private-key'
 
-      let tunnelConf = {
+      const tunnelConf = {
         username: 'bastion',
         host: 'bastion-host',
         privateKey: 'super-private-key',
@@ -163,79 +193,118 @@ describe('pg', () => {
         localPort: 49152
       }
 
-      let psql = require('../../lib/psql')
-      sinon.stub(psql, 'exec').returns(Promise.resolve(emptyResponse))
-      let cp = sinon.mock(require('child_process'))
-      let cmd = 'env PGSSLMODE=prefer pg_dump --verbose -F c -Z 0      localdb | env PGPASSWORD="pass" pg_restore --verbose --no-acl --no-owner -U jeff -h herokai.com -p 5432 -d mydb'
-      cp.expects('spawn').withExactArgs(cmd, [], opts).once().returns(
-        {
-          stdout: {
-            on: stdoutHandler
-          },
-          on: exitHandler
-        }
-      )
+      const dumpFlags = ['--verbose', '-F', 'c', '-Z', '0', 'localdb']
+      const restoreFlags = ['--verbose', '-F', 'c', '--no-acl', '--no-owner', '-U', 'jeff', '-h', 'herokai.com', '-p', '5432', '-d', 'mydb']
+
+      spawnStub.withArgs('pg_dump', dumpFlags, dumpOpts).returns({
+        stdout: {
+          pipe: () => {}
+        },
+        on: exitHandler
+      })
+      spawnStub.withArgs('pg_restore', restoreFlags, restoreOpts).returns({
+        stdin: {
+          end: () => {}
+        },
+        on: exitHandler
+      })
 
       return push.run({ args: { source: 'localdb', target: 'postgres-1' }, flags: {} })
-        .then(() => cp.verify())
+        .then(() => expect(spawnStub.callCount, 'to equal', 2))
         .then(() => expect(cli.stdout, 'to equal', 'heroku-cli: Pushing localdb ---> postgres-1\nheroku-cli: Pushing complete.\n'))
         .then(() => expect(cli.stderr, 'to equal', ''))
-        .then(() => expect(
-          tunnelStub.withArgs(tunnelConf).calledOnce, 'to equal', true)
-        )
-        .then(() => psql.exec.restore())
-    }))
+        .then(() => expect(tunnelStub.withArgs(tunnelConf).calledOnce, 'to equal', true))
+    })
 
-    it('exits non-zero when there is an error', sinon.test(() => {
-      let psql = require('../../lib/psql')
-      sinon.stub(psql, 'exec').returns(Promise.resolve(emptyResponse))
+    it('exits non-zero when there is an error', () => {
+      const dumpFlags = ['--verbose', '-F', 'c', '-Z', '0', 'localdb']
+      const restoreFlags = ['--verbose', '-F', 'c', '--no-acl', '--no-owner', '-U', 'jeff', '-h', 'herokai.com', '-p', '5432', '-d', 'mydb']
 
-      let cp = sinon.mock(require('child_process'))
-      let cmd = 'env PGSSLMODE=prefer pg_dump --verbose -F c -Z 0      localdb | env PGPASSWORD="pass" pg_restore --verbose --no-acl --no-owner -U jeff -h herokai.com -p 5432 -d mydb'
-      cp.expects('spawn').withExactArgs(cmd, [], opts).once().returns(
-        {
-          stdout: {
-            on: stdoutHandler
-          },
-          on: function (key, func) {
-            func(1)
-          }
+      spawnStub.withArgs('pg_dump', dumpFlags, dumpOpts).returns({
+        stdout: {
+          pipe: () => {}
+        },
+        on: (key, func) => {
+          func(1)
         }
-      )
+      })
+      spawnStub.withArgs('pg_restore', restoreFlags, restoreOpts).returns({
+        stdin: {
+          end: () => {}
+        }
+      })
 
-      return expect(push.run({ args: { source: 'localdb', target: 'postgres-1' }, flags: {} }), 'to be rejected with', { code: 1 })
-        .then(() => cp.verify())
-        .then(() => expect(cli.stdout, 'to equal', 'heroku-cli: Pushing localdb ---> postgres-1\n'))
-        .then(() => expect(cli.stderr, 'to equal', ''))
-        .then(() => psql.exec.restore())
-    }))
+      return push.run({ args: { source: 'localdb', target: 'postgres-1' }, flags: {} })
+        .catch((error) => {
+          expect(error.message, 'to equal', 'pg_dump errored with 1')
+          expect(spawnStub.callCount, 'to equal', 2)
+          expect(cli.stdout, 'to equal', 'heroku-cli: Pushing localdb ---> postgres-1\n')
+          expect(cli.stderr, 'to equal', '')
+        })
+    })
   })
 
   describe('pull', () => {
-    it('pulls a db in', sinon.test(() => {
-      let psql = require('../../lib/psql')
+    const dumpFlags = ['--verbose', '-F', 'c', '-Z', '0', '-U', 'jeff', '-h', 'herokai.com', '-p', '5432', 'mydb']
+    const dumpOpts = {
+      env: {
+        PGPASSWORD: 'pass',
+        PGSSLMODE: 'prefer',
+        ...env
+      },
+      stdio: ['pipe', 'pipe', 2],
+      encoding: 'utf8',
+      shell: true
+    }
+    const restoreFlags = ['--verbose', '-F', 'c', '--no-acl', '--no-owner', '-d', 'localdb']
+    const restoreOpts = {
+      env: { ...env },
+      stdio: ['pipe', 'pipe', 2],
+      encoding: 'utf8',
+      shell: true
+    }
+
+    const childProcess = require('child_process')
+    const psql = require('../../lib/psql')
+
+    let createDbStub
+    let spawnStub
+
+    beforeEach(() => {
+      createDbStub = sinon.stub(childProcess, 'execSync')
+      spawnStub = sinon.stub(childProcess, 'spawn')
       sinon.stub(psql, 'exec')
-      let cp = sinon.mock(require('child_process'))
-      cp.expects('execSync').withExactArgs('createdb     localdb', { stdio: 'inherit' }).once()
 
-      let cmd = 'env PGPASSWORD="pass" PGSSLMODE=prefer pg_dump --verbose -F c -Z 0  -U jeff -h herokai.com -p 5432  mydb | env pg_restore --verbose --no-acl --no-owner    -d localdb'
-      cp.expects('spawn').withExactArgs(cmd, [], opts).once().returns(
-        {
-          stdout: {
-            on: stdoutHandler
-          },
-          on: exitHandler
-        }
-      )
+      spawnStub.withArgs('pg_dump', dumpFlags, dumpOpts).returns({
+        stdout: {
+          pipe: () => {}
+        },
+        on: exitHandler
+      })
+      spawnStub.withArgs('pg_restore', restoreFlags, restoreOpts).returns({
+        stdin: {
+          end: () => {}
+        },
+        on: exitHandler
+      })
+    })
 
+    afterEach(() => {
+      psql.exec.restore()
+      createDbStub.restore()
+      spawnStub.restore()
+    })
+
+    it('pulls a db in', () => {
       return pull.run({ args: { source: 'postgres-1', target: 'localdb' }, flags: {} })
-        .then(() => cp.verify())
+        .then(() => expect(createDbStub.calledOnce, 'to equal', true))
+        .then(() => expect(createDbStub.calledWithExactly('createdb localdb', { stdio: 'inherit' }), 'to equal', true))
+        .then(() => expect(spawnStub.callCount, 'to equal', 2))
         .then(() => expect(cli.stdout, 'to equal', 'heroku-cli: Pulling postgres-1 ---> localdb\nheroku-cli: Pulling complete.\n'))
         .then(() => expect(cli.stderr, 'to equal', ''))
-        .then(() => psql.exec.restore())
-    }))
+    })
 
-    it('opens an SSH tunnel and runs pg_dump for bastion databases', sinon.test(() => {
+    it('opens an SSH tunnel and runs pg_dump for bastion databases', () => {
       db.bastionHost = 'bastion-host'
       db.bastionKey = 'super-private-key'
 
@@ -249,29 +318,13 @@ describe('pg', () => {
         localPort: 49152
       }
 
-      let psql = require('../../lib/psql')
-      sinon.stub(psql, 'exec')
-      let cp = sinon.mock(require('child_process'))
-      cp.expects('execSync').withExactArgs('createdb     localdb', { stdio: 'inherit' }).once()
-
-      let cmd = 'env PGPASSWORD="pass" PGSSLMODE=prefer pg_dump --verbose -F c -Z 0  -U jeff -h herokai.com -p 5432  mydb | env pg_restore --verbose --no-acl --no-owner    -d localdb'
-      cp.expects('spawn').withExactArgs(cmd, [], opts).once().returns(
-        {
-          stdout: {
-            on: stdoutHandler
-          },
-          on: exitHandler
-        }
-      )
-
       return pull.run({ args: { source: 'postgres-1', target: 'localdb' }, flags: {} })
-        .then(() => cp.verify())
+        .then(() => expect(createDbStub.calledOnce, 'to equal', true))
+        .then(() => expect(createDbStub.calledWithExactly('createdb localdb', { stdio: 'inherit' }), 'to equal', true))
+        .then(() => expect(spawnStub.callCount, 'to equal', 2))
         .then(() => expect(cli.stdout, 'to equal', 'heroku-cli: Pulling postgres-1 ---> localdb\nheroku-cli: Pulling complete.\n'))
         .then(() => expect(cli.stderr, 'to equal', ''))
-        .then(() => expect(
-          tunnelStub.withArgs(tunnelConf).calledOnce, 'to equal', true)
-        )
-        .then(() => psql.exec.restore())
-    }))
+        .then(() => expect(tunnelStub.withArgs(tunnelConf).calledOnce, 'to equal', true))
+    })
   })
 })