+++ packages/autocomplete/CHANGELOG.md
@@ -3,6 +3,63 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.18.0](https://github.com/heroku/cli/compare/v7.17.2...v7.18.0) (2018-10-17)
+
+**Note:** Version bump only for package @heroku-cli/plugin-autocomplete
+
+
+
+
+
+## [7.17.2](https://github.com/heroku/cli/compare/v7.17.1...v7.17.2) (2018-10-16)
+
+
+### Bug Fixes
+
+* **autocomplete:** skip updating resource completions if not logged in ([#1042](https://github.com/heroku/cli/issues/1042)) ([cb09499](https://github.com/heroku/cli/commit/cb09499))
+
+
+
+
+
+## [7.17.1](https://github.com/heroku/cli/compare/v7.17.0...v7.17.1) (2018-10-16)
+
+**Note:** Version bump only for package @heroku-cli/plugin-autocomplete
+
+
+
+
+
+## [7.16.7](https://github.com/heroku/cli/compare/v7.16.6...v7.16.7) (2018-10-10)
+
+
+### Bug Fixes
+
+* updated command ([fbd11fd](https://github.com/heroku/cli/commit/fbd11fd))
+
+
+
+
+
+## [7.16.6](https://github.com/heroku/cli/compare/v7.16.5...v7.16.6) (2018-10-05)
+
+**Note:** Version bump only for package @heroku-cli/plugin-autocomplete
+
+
+
+
+
+<a name="7.16.3"></a>
+## [7.16.3](https://github.com/heroku/cli/compare/v7.16.2...v7.16.3) (2018-10-03)
+
+
+### Bug Fixes
+
+* updated deps ([#1058](https://github.com/heroku/cli/issues/1058)) ([ce1fd6b](https://github.com/heroku/cli/commit/ce1fd6b))
+
+
+
+
 <a name="7.16.2"></a>
 ## [7.16.2](https://github.com/heroku/cli/compare/v7.16.1...v7.16.2) (2018-10-02)