+++ packages/local-v5/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.18.0](https://github.com/heroku/cli/compare/v7.17.2...v7.18.0) (2018-10-17)
+
+**Note:** Version bump only for package @heroku-cli/plugin-local-v5
+
+
+
+
+
+## [7.16.7](https://github.com/heroku/cli/compare/v7.16.6...v7.16.7) (2018-10-10)
+
+
+### Bug Fixes
+
+* updated command ([fbd11fd](https://github.com/heroku/cli/commit/fbd11fd))
+
+
+
+
+
 <a name="7.16.2"></a>
 ## [7.16.2](https://github.com/heroku/cli/compare/v7.16.1...v7.16.2) (2018-10-02)