+++ packages/pg-v5/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.18.0](https://github.com/heroku/cli/compare/v7.17.2...v7.18.0) (2018-10-17)
+
+
+### Bug Fixes
+
+* pass env vars down to pg_pull/pg_drop ([#1074](https://github.com/heroku/cli/issues/1074)) ([5442383](https://github.com/heroku/cli/commit/5442383))
+
+
+### Features
+
+* **pg:** add psql history env var ([28e74a8](https://github.com/heroku/cli/commit/28e74a8))
+
+
+
+
+
+## [7.17.1](https://github.com/heroku/cli/compare/v7.17.0...v7.17.1) (2018-10-16)
+
+
+### Bug Fixes
+
+* allow fetching postgres bastion keys via data api ([#1068](https://github.com/heroku/cli/issues/1068)) ([bec6d61](https://github.com/heroku/cli/commit/bec6d61))
+
+
+
+
+
+## [7.16.8](https://github.com/heroku/cli/compare/v7.16.7...v7.16.8) (2018-10-11)
+
+
+### Bug Fixes
+
+* **pg:** push and pull for windows ([66587d8](https://github.com/heroku/cli/commit/66587d8))
+
+
+
+
+
 <a name="7.16.2"></a>
 ## [7.16.2](https://github.com/heroku/cli/compare/v7.16.1...v7.16.2) (2018-10-02)